+++ src/OperatorOverloads/Binary.cpp
@@ -8,7 +8,7 @@ BigInteger BigInteger::operator+(const BigInteger& rhs) const {
 	return ret;
 }
 
-BigInteger BigInteger::operator+(const unsigned long long& rhs) const {
+BigInteger BigInteger::operator+(const uint64_t& rhs) const {
 	BigInteger ret = BigInteger(*this);
 	ret += rhs;
 	return ret;
@@ -22,7 +22,7 @@ BigInteger BigInteger::operator-(const BigInteger& rhs) const {
 	return ret;
 }
 
-BigInteger BigInteger::operator-(const unsigned long long& rhs) const {
+BigInteger BigInteger::operator-(const uint64_t& rhs) const {
 	BigInteger ret = BigInteger(*this);
 	ret -= rhs;
 	return ret;
@@ -36,7 +36,7 @@ BigInteger BigInteger::operator*(const BigInteger& rhs) const {
 	return ret;
 }
 
-BigInteger BigInteger::operator*(const unsigned long long& rhs) const {
+BigInteger BigInteger::operator*(const uint64_t& rhs) const {
 	BigInteger ret = BigInteger(*this);
 	ret *= rhs;
 	return ret;
@@ -50,7 +50,7 @@ BigInteger BigInteger::operator/(const BigInteger& rhs) const {
 	return ret;
 }
 
-BigInteger BigInteger::operator/(const unsigned long long& rhs) const {
+BigInteger BigInteger::operator/(const uint64_t& rhs) const {
 	BigInteger ret = BigInteger(*this);
 	ret /= rhs;
 	return ret;
@@ -64,7 +64,7 @@ BigInteger BigInteger::operator%(const BigInteger& rhs) const {
 	return ret;
 }
 
-BigInteger BigInteger::operator%(const unsigned long long& rhs) const {
+BigInteger BigInteger::operator%(const uint64_t& rhs) const {
 	BigInteger ret = BigInteger(*this);
 	ret %= rhs;
 	return ret;
